+++ env/qunex_env_status.sh
@@ -210,13 +210,9 @@ if [[ "$1" == "--envstatus" ]] || [[ "$1" == "--envreport" ]] || [[ "$1" == "--e
     unset BinaryError
 
     ## -- Check for HCPpipedir
-    if [[ -e $HCPPIPEDIR/global/scripts/versioning/base.txt ]]; then
+    if [[ -e $HCPPIPEDIR/show_version ]]; then
         # add specific TAG and commit hash
-        echo "    HCPpipelines TAG : $(cat $HCPPIPEDIR/global/scripts/versioning/base.txt)"
-        echo " HCPpipelines commit : $(git --git-dir ${HCPPIPEDIR}/.git log -1 --pretty=format:"%H")"
-    elif [[ -e $HCPPIPEDIR/version.txt ]]; then
-        # add specific TAG and commit hash
-        echo "    HCPpipelines TAG : $(cat $HCPPIPEDIR/version.txt)"
+        echo "    HCPpipelines TAG : `$HCPPIPEDIR/show_version --short`"
         echo " HCPpipelines commit : $(git --git-dir ${HCPPIPEDIR}/.git log -1 --pretty=format:"%H")"
     else
         BinaryError="yes"; BinaryErrorReport="HCPPipelines"